Buffer fix & Standardization of ID
Buffer fix: fixed the buffer addresses. Now start at 512 and has 511 bytes.

Standardization of ID:
This pack will now have hispeed ID's that can be used to identificate the device. This is because it's an unofficial wiremod pack and External Bus cannot identify the device. It will return 1 as ID. So now you can look for address 511 for the correct ID.